Refactor findRepresentativeLabel functions into utils
